connected add method to add button
A check for a null connection will be required for each db methid, in case
the connection attempt failed in the constructor.

Also a run method from the form will be required to be called as worker fo
rthe thread. Then it can handle message boxes indicating the success or
failure of th eopperation.